+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - tidy3d improvements:
     - get_simulation and write_sparameters accepts componentOrFactory
+    - grating_coupler simulations can also be dispersive
 
 ## [4.3.3](https://github.com/gdsfactory/gdsfactory/pull/214)